% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{abort_bad_argument}
\alias{abort_bad_argument}
\title{checks arguments types}
\usage{
abort_bad_argument(arg, must, not = NULL, call = rlang::caller_env())
}
\arguments{
\item{call}{}
}
\description{
checks arguments types
}
