% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{abort_not_found}
\alias{abort_not_found}
\title{Check if file exists and is not empty}
\usage{
abort_not_found(x, call = rlang::caller_env())
}
\arguments{
\item{x}{}
}
\description{
Check if file exists and is not empty
}
