% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCurve-package.R
\docType{package}
\name{batchCurve-package}
\alias{batchCurve}
\alias{batchCurve-package}
\title{batchCurve: Provides batch analysis for dose-response modelling}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A tailored package to analyse raw dose-response data and corresponding meta data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rayVonW/batchCurve}
  \item \url{https://rayvonw.github.io/batchCurve/}
  \item Report bugs at \url{https://github.com/rayVonW/batchCurve/issues}
}

}
\keyword{internal}
