% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{batchCurve_example}
\alias{batchCurve_example}
\title{Get path to readr example}
\usage{
batchCurve_example(file = NULL)
}
\arguments{
\item{file}{Name of file. If `NULL`, the example files will be listed.}
}
\description{
batchCurve comes bundled with a number of sample files in its `inst/extdata`
directory. This function make them easy to access
}
\examples{
readr_example()
readr_example("TRno4850.csv")
}
