% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planner.R
\name{build_plates}
\alias{build_plates}
\title{Converts assay parameters to 384w plate layouts in excel for Tecan D300e printing}
\usage{
build_plates(data, prefix = "example")
}
\arguments{
\item{data}{A data frame of Compound dilution curve information.}

\item{prefix}{A text string for output file names.}
}
\value{
An export .xlsx file with each sheet representing a 384w plate for Tecan D300e.
}
\description{
Converts assay parameters to 384w plate layouts in excel for Tecan D300e printing
}
