% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_if_not_numeric}
\alias{check_if_not_numeric}
\title{Checks if numeric}
\usage{
check_if_not_numeric(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{}
}
\description{
Checks if numeric
}
