% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{check_raw_files}
\alias{check_raw_files}
\title{Checks for the existence of raw plate reader files, prefixed with either 'TRno' or 'automated'}
\usage{
check_raw_files(file_path)
}
\arguments{
\item{file_path}{A path to the directory where raw data is stored with  meta file}
}
\value{
A list of raw files for analysis.
}
\description{
Checks for the existence of raw plate reader files, prefixed with either 'TRno' or 'automated'
}
