% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Wilcox rank sum of unpaired IC50 data}
\usage{
compare(data, prefix = "group")
}
\arguments{
\item{data}{A data frame - exported object from fit_data().}

\item{prefix}{A string for output file names.}
}
\value{
A data frame object of mean IC50 data and p values
}
\description{
Wilcox rank sum of unpaired IC50 data
}
