% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_meta}
\alias{example_meta}
\title{Dose response meta data example}
\format{
## `example_meta`
A data frame with 23 rows and 9 columns:
\describe{
  \item{plate_id}{ID given to each plates reader file}
  \item{position_id}{Position of an assay on plate, 'pos_x'}
  \item{format}{Plate format, either '384w' or '96w'}
  \item{replicates}{How many replicates per assay used, either '2' or '3'}
  \item{index}{label for compound, used for sorting output}
  \item{compound}{Supplier ID of compound}
  \item{cell}{Cell ID}
  \item{starting_uM}{Drug concentration in column 1 of plate, i.e. the highest concentration in uM}
  \item{dilution_factor}{The serial dilution factor used in the assay}
  ...
}
}
\source{
author
}
\usage{
data(example_meta)
}
\description{
An example meta data file describing drug assays for analysis ...
}
\keyword{dataset}
