% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{fit.LL4}
\alias{fit.LL4}
\title{drc fitting function}
\usage{
fit.LL4(data)
}
\arguments{
\item{data}{Normalised measurements from a drug assay.}
}
\value{
An LL4 fitted model or error to handle
}
\description{
drc fitting function
}
