% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_caller.R
\name{fit_data}
\alias{fit_data}
\title{iterates over assays for IC50 estimation}
\usage{
fit_data(file_path, batch_id)
}
\arguments{
\item{file_path}{A file containing the batches meta data}

\item{batch_id}{A unique character value to identify the batch}
}
\value{
A data frame of LL4 model coefficients for dose-response assays
}
\description{
iterates over assays for IC50 estimation
}
