% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma-package.R
\docType{package}
\name{gamma-package}
\alias{gamma}
\alias{gamma-package}
\title{gamma: Provides batch analysis for dose-response modelling}
\description{
A tailored package to analyse raw dose-response data and corresponding meta data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rayVonW/gamma}
  \item \url{https://rayvonw.github.io/gamma/}
  \item Report bugs at \url{https://github.com/rayVonW/gamma/issues}
}

}
\keyword{internal}
