% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{generate_curve}
\alias{generate_curve}
\title{Generate curves from LL4 coefficients for drug assay plots.}
\usage{
generate_curve(data)
}
\arguments{
\item{data}{A data frame of model coefficients and drug assay data.}
}
\value{
A data frame of predicted values using the model and dose range.
}
\description{
Generate curves from LL4 coefficients for drug assay plots.
}
