% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_layout.R
\name{get_locations}
\alias{get_locations}
\title{finds drug assay data on a 96 or 384 well plate}
\usage{
get_locations(assay, plate)
}
\arguments{
\item{plate}{}
}
\description{
finds drug assay data on a 96 or 384 well plate
}
