% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{import_plate}
\alias{import_plate}
\title{checks raw data files}
\usage{
import_plate(file, meta)
}
\arguments{
\item{file}{List of raw data files matching prefix}

\item{meta}{The meta dataframe.}
}
\value{
A list of plate data: 1. raw data, 2. plate ID, 3.date 4. filename
}
\description{
checks raw data files
}
