% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planner.R
\name{plan_ranges}
\alias{plan_ranges}
\title{Designs dose-response assays from 96 well compounds plate for Tecan D300e}
\usage{
plan_ranges(input_file)
}
\arguments{
\item{input_file}{A file name containing Compound assay parameters or guide IC50 values.}
}
\value{
A data frame of completed dose ranges.
}
\description{
Designs dose-response assays from 96 well compounds plate for Tecan D300e
}
