% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_caller.R
\name{plot_fit}
\alias{plot_fit}
\title{Produces group plots of dose response models from fit_data()}
\usage{
plot_fit(
  results,
  data,
  prefix,
  plot.var = "compound",
  colour.var = "cell",
  facet.var = NULL,
  grid.var = 3,
  leg.pos = "bottom"
)
}
\arguments{
\item{results}{A data frame, from fit_data.}

\item{data}{A data frame, from fit_data.}

\item{prefix}{An character ID for output file prefix.}

\item{plot.var}{A column name you wish to subset data per plot, default is "compound", you may wish specify "cell".}

\item{colour.var}{A column name you wish to differentiate data per plot by colour, default = "cell", you may wish specify "compound".}

\item{facet.var}{A column name you wish to differentiate data by facet, default = NULL, you may wish to have "cell" on separate plots.}

\item{grid.var}{a numeric value specifying the grid size in the exported pdf, default = 3 i.e 3x3.}

\item{leg.pos}{Legend position accepts 'right', 'left' top'}
}
\value{
A list of plot objects per plot_var
}
\description{
Produces group plots of dose response models from fit_data()
}
