% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{plot_mean}
\alias{plot_mean}
\title{Plots barchart of mean I50 values and displays p value}
\usage{
plot_mean(data, prefix = "group", grid.var = 2)
}
\arguments{
\item{data}{A data frame direct output from compare()}

\item{prefix}{A string values to append output files}

\item{grid.var}{An integer value describing the plot grid size in output file}
}
\value{
A list of plot objects
}
\description{
Plots barchart of mean I50 values and displays p value
}
