% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{retrieve_results}
\alias{retrieve_results}
\title{adds model coefficients to data}
\usage{
retrieve_results(model, assays)
}
\arguments{
\item{model}{LL4 model fitted to the data.}

\item{assays}{Assay meta data}
}
\value{
Assay meta dataframe with LL4 coefficiants appended
}
\description{
adds model coefficients to data
}
