% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_caller.R
\name{run}
\alias{run}
\title{iterates over assays for IC50 estimation}
\usage{
run(file_path, batch_id)
}
\arguments{
\item{file_path}{A file containing the batches meta data}

\item{batch_id}{A unique character value to identify the batch}
}
\value{
A dataframe of LL4 model coefficiants for dose-response assays
}
\description{
iterates over assays for IC50 estimation
}
