% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test.fun}
\alias{test.fun}
\title{embedded utility function}
\usage{
test.fun(a, b)
}
\arguments{
\item{a}{a number}

\item{b}{a number}
}
\description{
embedded utility function
}
\examples{
test.fun(1+1)
}
