% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{test.src}
\alias{test.src}
\title{tests functions in other scripts}
\usage{
test.src(c, d)
}
\arguments{
\item{c}{A number.}

\item{d}{A number.}
}
\description{
tests functions in other scripts
}
\examples{
test.src(1,1)


}
