% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{validate_meta}
\alias{validate_meta}
\title{validates the meta file for assays and summarise contents}
\usage{
validate_meta(df)
}
\arguments{
\item{df}{A data frame of user supplied meta csv data.}
}
\value{
A data frame of user supplied assay meta data.
}
\description{
validates the meta file for assays and summarise contents
}
