% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{fit.LL4}
\alias{fit.LL4}
\title{drc fitting function}
\usage{
fit.LL4(data, assay)
}
\arguments{
\item{data}{Normalised measurements from a drug assay.}

\item{assay}{Assay meta data.}
}
\value{
An LL4 fitted model
}
\description{
drc fitting function
}
