% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_layout.R
\name{get_assay_data}
\alias{get_assay_data}
\title{subsets plate data per assay, converts values to percentages then adds the dose curve}
\usage{
get_assay_data(data, location, assays_info)
}
\arguments{
\item{assays_info}{}
}
\description{
subsets plate data per assay, converts values to percentages then adds the dose curve
}
