% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_csv_file}
\alias{read_csv_file}
\title{reads in csv of meta data for drug assays}
\usage{
read_csv_file(file_path)
}
\arguments{
\item{file_path}{}
}
\value{
A dataframe
}
\description{
reads in csv of meta data for drug assays
}
