% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{retrieve_results}
\alias{retrieve_results}
\title{adds model coefficients to data}
\usage{
retrieve_results(model, assays, locate)
}
\arguments{
\item{model}{An LL4 model fitted to the data.}

\item{assays}{A data frame of assay meta data.}

\item{locate}{A data frame of coordinates.}
}
\value{
Assay meta data frame with LL4 coefficients appended
}
\description{
adds model coefficients to data
}
