% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{validate_meta}
\alias{validate_meta}
\title{validates the meta file for assays and summarise contents}
\usage{
validate_meta(df)
}
\arguments{
\item{df}{from the user csv file}
}
\value{
a dataframe
}
\description{
validates the meta file for assays and summarise contents
}
